Added hotfix to force https due to https://github.com/pkp/pkp-lib/issues/1504 All this does is change getProtocol to always return 'https' (see lines 370-373 of PKPRequest.inc.php).
This is because we currently run OJS Behind a proxy that does ssl termination so this is a bit of a problem otherwise.

Probably should submit a PR to pkp-lib but will take time to be accepted and will then need to be merged into OJS so this is the easiest fix for now.

Setting force_ssl in config results in inf redirects as pkp looks at the apache refer details and tries to redirect you.
Configuring 'allowProtocolRelative' for frontend asset calls seems like a better way to go but this touches a lot of files from what I can see.