Fix sed to replace the whole line for DB name
Previously, `sed` was replacing only a portion of the line
containing the DB name. This resulted in a part of the
DB name being left over each time the container was restarted,
leaving a string like 'name = ojs-production-production-production'.

Now, we are using sed to replace the contents of the whole line,
which should fix a problem we're having with the container being
able to find the MySQL DB.